refactor: Organizando Pacotes, removendo 'data' consumindo api em um mesmo pacote, atualizando DTOs
