+++ app/src/main/java/com/example/mobilevan/config/RetrofitConfig.kt
@@ -1,5 +1,6 @@
 package com.example.mobilevan.config
 
+import com.example.mobilevan.service.WeatherApi
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
@@ -11,4 +12,17 @@ object RetrofitConfig {
         .client(OkHttpClient())
         .addConverterFactory(GsonConverterFactory.create())
         .build()
+}
+
+object RetrofitConfigHgbr {
+
+    private const val BASE_URL = "https://api.hgbrasil.com/"
+
+    val api: WeatherApi by lazy {
+        Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+            .create(WeatherApi::class.java)
+    }
 }
