--- app/src/main/java/com/example/mobilevan/data/remote/RetrofitInstance.kt
@@ -1,17 +1,0 @@
-package com.example.mobilevan.data.remote
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object RetrofitInstance {
-
-    private const val BASE_URL = "https://api.hgbrasil.com/"
-
-    val api: WeatherApi by lazy {
-        Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(WeatherApi::class.java)
-    }
-}
