+++ app/src/main/java/com/example/mobilevan/service/dto/ForecastDay.kt
@@ -1,0 +1,11 @@
+package com.example.mobilevan.service.dto
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ForecastDay(
+    val weekday: String,
+    val date: String,
+    val max: Int,
+    val min: Int,
+    val description: String
+)
