+++ app/src/main/java/com/example/mobilevan/service/dto/WeatherResponse.kt
@@ -1,0 +1,9 @@
+package com.example.mobilevan.service.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class WeatherResponse(
+    @SerialName("results") val results: WeatherResults
+)
