+++ app/src/main/java/com/example/mobilevan/service/dto/WeatherResults.kt
@@ -1,0 +1,12 @@
+package com.example.mobilevan.service.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class WeatherResults(
+    val temp: Int,
+    val city: String,
+    @SerialName("date") val date: String,
+    @SerialName("forecast") val forecast: List<ForecastDay>
+)
