--- app/src/main/java/com/example/mobilevan/ui/screens/feature_clima/MainViewModel.kt
@@ -1,2 +1,0 @@
-package com.example.mobilevan.ui.screens.feature_clima
-
